drive: optimize performance for --min-age, --max-age (#5023)
The Google Drive API allows to place a clause like "modifiedTime > '2012-06-04T12:00:00'"
in the query param, so the filterflags --max-age and --min-age can be applied directly
at the directory listing phase rather than in a filter.
This is extremely helpful when we want to do an incremental backup of a remote drive
with a large number of files but the number of recently changed file is small.

Co-authored-by: fotile96 <fotile96@users.noreply.github.com>